requirements.txt
APScheduler==3.11.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
openpyxl==3.1.5
packaging==25.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
